requirements.txt
google-generativeai
python-dotenv
PyPDF2
textstat
requests
beautifulsoup4
lxml
sentence-transformers
numpy<2.0
reportlab
shiny
pandas
plotly
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
